billing-api/requirements.txt
Babel==2.3.4
bcrypt==3.1.1
cffi==1.8.3
click==6.6
dateutils==0.6.6
debtcollector==1.8.0
docopt==0.6.2
Flask==0.11.1
Flask-Login==0.3.2
funcsigs==1.0.2
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
keystoneauth1==2.12.1
MarkupSafe==0.23
monotonic==1.2
msgpack-python==0.4.8
MySQL-python==1.2.5
netaddr==0.7.18
netifaces==0.10.5
oslo.config==3.17.0
oslo.i18n==3.9.0
oslo.serialization==2.13.0
oslo.utils==3.16.0
pbr==1.10.0
positional==1.1.1
pycparser==2.14
pyparsing==2.1.9
python-dateutil==2.5.3
python-keystoneclient==3.5.0
pytz==2016.7
records==0.4.3
requests==2.11.1
rfc3986==0.4.1
six==1.10.0
SQLAlchemy==1.0.15
stevedore==1.17.1
tablib==0.11.2
Werkzeug==0.11.11
wrapt==1.10.8
